requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Markdown==3.7
